Download binaries required for offline installations (#21)
This PR adds an action plugin which parses all of the role defaults variables and returns a list of all binaries required during the execution of the roles in the collection.
A playbook is included that uses this action plugin to automatically download all of the required binaries into a directory of the users choice.